requirements.txt
flask==3.0.2
requests==2.31.0
icecream==2.1.3
gitpython==3.1.27